refactor(desktop): isolate terminal session management
